requirements.txt
Django~=2.2.10
django-allauth~=0.41.0
django-widget-tweaks~=1.4.8
# stripe~=2.27.0
# pillow~=2.2.1
stripe~=3.5.0
pillow~=9.5.0